Update NSAFanalysis.R
Addition of Longitudinal analysis method with mixed model using Limma. 
Also it has been added the possibility to ignore groups with less than 3 samples during the filtering process, so the presence or absence in these group won't be take into account to filter out or not a protein, their samples for the selected proteins will be included with their abundance numbers.